+++ CHANGELOG.md
@@ -7,15 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2022-07-30
+
 ### Added
 
 - Update tasks.
 - GitHub Pages config.
 - Dependabot config.
+- PowerShell module with install command.
 
 ### Fixed
 
 - Dotnet-Matrix.Tasks does not fail in empty project.
+- PowerShell-Matrix.Tasks does not fail in empty project.
 - Wrong file separators for linux.
 
 ## [1.1.0] - 2022-03-19